graph_generation/requirements.txt
networkx
graphml
